>[project:amiya.akn,project:lungmen.akn,project:hass]: Rename pulumi src/ to stack/
Aligns directory naming with chezmoi.sh convention where
Pulumi stack definitions live under stack/ rather than src/.
