+++ projects/amiya.akn/src/infrastructure/pulumi/index.ts
@@ -1,6 +1,6 @@
-export * from "./src/argotails";
-export * from "./src/cert-manager";
-export * from "./src/cloudflare-operator";
-export * from "./src/pocket-id";
-export * from "./src/tailscale-operator";
-export * from "./src/vault";
+export * from "./stack/argotails";
+export * from "./stack/cert-manager";
+export * from "./stack/cloudflare-operator";
+export * from "./stack/pocket-id";
+export * from "./stack/tailscale-operator";
+export * from "./stack/vault";
