+++ projects/amiya.akn/src/infrastructure/pulumi/stack/tailscale-operator.ts
