+++ projects/amiya.akn/src/infrastructure/pulumi/stack/vault.ts
