+++ projects/hass/src/infrastructure/pulumi/stack/home-assistant.ts
