+++ projects/lungmen.akn/src/infrastructure/pulumi/index.ts
@@ -1,3 +1,3 @@
-export * from "./src/cert-manager";
-export * from "./src/tailscale-operator";
-export * from "./src/vault";
+export * from "./stack/cert-manager";
+export * from "./stack/tailscale-operator";
+export * from "./stack/vault";
